Create model package and rename HistoryItem to LunchOrder
