+++ .idea/misc.xml
@@ -44,7 +44,10 @@
         <entry key="..\:/Universidad/Semestre8/AppsMoviles/LunchTicket/app/src/main/res/menu/restaurant_menu.xml" value="0.25416666666666665" />
         <entry key="..\:/Universidad/Semestre8/AppsMoviles/LunchTicket/app/src/main/res/menu/student_menu.xml" value="0.25416666666666665" />
         <entry key="..\:/Universidad/Semestre8/AppsMoviles/LunchTicket/app/src/main/res/xml/paths.xml" value="0.25416666666666665" />
+        <entry key="..\:/Universidad/Semestre8/AppsMoviles/TrabajoFinalVersiones/LunchTicket/app/src/main/res/drawable/edittext_background.xml" value="0.24259259259259258" />
         <entry key="..\:/Universidad/Semestre8/AppsMoviles/TrabajoFinalVersiones/LunchTicket/app/src/main/res/drawable/historyitem_background.xml" value="0.7928571428571428" />
+        <entry key="..\:/Universidad/Semestre8/AppsMoviles/TrabajoFinalVersiones/LunchTicket/app/src/main/res/drawable/iconimagebutton_background.xml" value="0.24259259259259258" />
+        <entry key="..\:/Universidad/Semestre8/AppsMoviles/TrabajoFinalVersiones/LunchTicket/app/src/main/res/drawable/mainbutton_background.xml" value="0.24259259259259258" />
         <entry key="..\:/Universidad/Semestre8/AppsMoviles/TrabajoFinalVersiones/LunchTicket/app/src/main/res/layout/activity_post_list.xml" value="0.15217391304347827" />
         <entry key="..\:/Universidad/Semestre8/AppsMoviles/TrabajoFinalVersiones/LunchTicket/app/src/main/res/layout/fragment_check_student.xml" value="0.15217391304347827" />
         <entry key="..\:/Universidad/Semestre8/AppsMoviles/TrabajoFinalVersiones/LunchTicket/app/src/main/res/layout/fragment_financial_profile.xml" value="0.264" />
