+++ app/src/main/java/com/example/lunchticket/model/Post.kt
@@ -1,4 +1,4 @@
-package com.example.lunchticket
+package com.example.lunchticket.model
 
 import java.io.Serializable
