+++ app/src/main/res/drawable/iconimagebutton_background.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape android:shape="rectangle"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/green"/>
+    <corners android:radius="10dp"/>
+</shape>
