+++ app/src/main/res/layout/fragment_restaurant_history.xml
@@ -72,11 +72,11 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView53" />
+        app:layout_constraintTop_toBottomOf="@+id/historyTotalItemsTV" />
 
     <TextView
-        android:id="@+id/textView53"
-        style="@style/SubTitleGray"
+        android:id="@+id/historyTotalItemsTV"
+        style="@style/DescriptionBold"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
