+++ app/src/main/res/values-night/themes.xml
@@ -207,7 +207,7 @@
     </style>
 
     <style name="ImageButtonGreen" parent="Widget.AppCompat.ImageButton">
-        <item name="android:background">@drawable/mainbutton_background</item>
+        <item name="android:background">@drawable/iconimagebutton_background</item>
         <item name="android:elevation">5dp</item>
     </style>
